Move the Specberus Wrapper to the lib folder
