+++ lib/orchestrator.js
@@ -9,7 +9,7 @@ var Promise = require('promise');
 var DocumentDownloader = require('./document-downloader');
 var JsonHttpService = require('./json-http-service');
 var Publisher = require('./publisher');
-var SpecberusWrapper = require('../functions.js').SpecberusWrapper;
+var SpecberusWrapper = require('./specberus-wrapper');
 var ThirdPartyChecker = require('./third-party-resources-checker');
 var TokenChecker = require('./token-checker').TokenChecker;
