+++ lib/specberus-wrapper.js
@@ -5,7 +5,7 @@ var List = require('immutable').List;
 var Map = require('immutable').Map;
 var Specberus = require('specberus/lib/validator').Specberus;
 
-require('./config.js');
+require('../config.js');
 
 var SpecberusWrapper = {};
 
@@ -59,4 +59,4 @@ SpecberusWrapper.validate = function (url) {
 
 SpecberusWrapper.version = new Specberus().version;
 
-exports.SpecberusWrapper = SpecberusWrapper;
+module.exports = SpecberusWrapper;
